refactor(api): extract collections from client.ts
